+++ sql-server-gen.cabal
@@ -51,19 +51,19 @@ cabal-version:       >=1.10
 
 library
   -- Modules exported by the library.
-  exposed-modules:     Database.SqlServer.Types.Database,
-                       Database.SqlServer.Types.Sequence,
-                       Database.SqlServer.Types.Identifiers,
-                       Database.SqlServer.Types.Procedure,
-                       Database.SqlServer.Types.Queue,
-                       Database.SqlServer.Types.DataTypes,
-                       Database.SqlServer.Types.Table,
-                       Database.SqlServer.Types.Collations,
-                       Database.SqlServer.Types.Certificate,
-                       Database.SqlServer.Types.User,
-                       Database.SqlServer.Types.FullTextCatalog,
-                       Database.SqlServer.Types.FullTextStopList,
-                       Database.SqlServer.Types.Login
+  exposed-modules:     Database.SqlServer.Definitions.Database,
+                       Database.SqlServer.Definitions.Sequence,
+                       Database.SqlServer.Definitions.Identifiers,
+                       Database.SqlServer.Definitions.Procedure,
+                       Database.SqlServer.Definitions.Queue,
+                       Database.SqlServer.Definitions.DataTypes,
+                       Database.SqlServer.Definitions.Table,
+                       Database.SqlServer.Definitions.Collations,
+                       Database.SqlServer.Definitions.Certificate,
+                       Database.SqlServer.Definitions.User,
+                       Database.SqlServer.Definitions.FullTextCatalog,
+                       Database.SqlServer.Definitions.FullTextStopList,
+                       Database.SqlServer.Definitions.Login
 
   ghc-options:         -Wall
