+++ src/Database/SqlServer/Definitions/Credential.hs
@@ -2,10 +2,10 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE GADTs #-}
 
-module Database.SqlServer.Types.Credential where
+module Database.SqlServer.Definitions.Credential where
 
-import Database.SqlServer.Types.Identifiers hiding (unwrap)
-import Database.SqlServer.Types.Entity
+import Database.SqlServer.Definitions.Identifiers hiding (unwrap)
+import Database.SqlServer.Definitions.Entity
 
 import Text.PrettyPrint
 import Test.QuickCheck
